requirements.txt
python-telegram-bot
python-dotenv
langchain
openai
pydub